Reorganize types: rename to kebab-case, add @file headers, move token-layout types
- emptyFactories.ts → empty-factories.ts, typeGuards.ts → type-guards.ts
- Extract FocusContext/SlotFocusInfo/TokenGroup/LinkSlot/RenderUnit from
  utils/token-layout.ts into types/token-layout.ts
- Add @file JSDoc to all three type files
- Update all import sites (20 files)
