+++ src/components/modals/CreateProjectModal.tsx
@@ -3,7 +3,7 @@ import { useLocalizedStrings } from '@papi/frontend/react';
 import { Button } from 'platform-bible-react';
 import { useState, useCallback, useRef } from 'react';
 import type { InterlinearProjectSummary } from '../../types/interlinear-project-summary';
-import { isInterlinearProjectSummary } from '../../types/typeGuards';
+import { isInterlinearProjectSummary } from '../../types/type-guards';
 
 /** Localized string keys used by {@link CreateProjectModal}. */
 const CREATE_PROJECT_MODAL_STRING_KEYS: `%${string}%`[] = [
